Fix: Move opendata to webcontroller, change title to .env APP_NAME
