+++ routes/api.php
@@ -2,7 +2,7 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\OpenDataController;
+use App\Http\Controllers\WebController;
 
 /*
 |--------------------------------------------------------------------------
@@ -18,4 +18,4 @@ use App\Http\Controllers\OpenDataController;
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
-Route::get('/data-kerjasama', [OpenDataController::class, 'getKerjasamaStats']);
+Route::get('/data-kerjasama', [WebController::class, 'getKerjasamaStats']);
